Added metadata serialization to contain token, added datetime serialization which was not implemented.
Added a default admin user in the db which is created on the first run of the server and has the username 'admin' and password 'admin'.

Added authenticate request handling to the server, which checks the username and password and returns a token if the user is authenticated (delegated auth service).

Modified request methods in the CLI to include the token.

Added a new "authenticate" command to the CLI to authenticate the user and store the token in a file.